requirement.txt
Django (1.8)
django-cors-headers (1.1.0)
django-filter (0.11.0)
djangorestframework (3.3.2)
Markdown (2.6.5)
pip (7.1.2)
setuptools (18.2)
wheel (0.24.0)
whitenoise (2.0)
